+++ build_vim.sh
@@ -1,19 +1,5 @@
 #!/usr/bin/env bash
 
-if [ "$(uname -s)" == "Darwin" ]; then
-  export NCPUS=`sysctl -n hw.ncpu`
-  echo "Not tested on MacOS. But why not use Homebrew? Exiting..."
-  exit 1
-elif [ "$(uname -s)" == "Linux" ]; then
-  export NCPUS=`nproc`
-fi
-
-# TODO: user-configurable source and install directories
-CLONE_DIR=$HOME/devel
-INSTALL_DIR=$HOME/local
-
-echo "Cloning to ${REPO_DIR}, and installing to ${INSTALL_DIR}..."
-
 # Prerequisites, according to
 # https://github.com/Valloric/YouCompleteMe/wiki/Building-Vim-from-source
 #
@@ -22,10 +8,48 @@ echo "Cloning to ${REPO_DIR}, and installing to ${INSTALL_DIR}..."
 #    libcairo2-dev libx11-dev libxpm-dev libxt-dev python-dev \
 #    python3-dev ruby-dev lua5.1 lua5.1-dev libperl-dev git
 
+if [[ "$(uname -s)" == "Darwin" ]]; then
+  export NCPUS=`sysctl -n hw.ncpu`
+  echo "Not tested on MacOS. Homebrew provides an up-to-date version of vim."
+  echo "Exiting..."
+  exit 1
+elif [[ "$(uname -s)" == "Linux" ]]; then
+  export NCPUS=`nproc`
+fi
+
+print_help()
+{
+  echo ""
+  echo "Usage:"
+  echo "  build_vim.sh -s [SOURCE_DIR] -t [INSTALL_DIR]"
+  echo "where SOURCE_DIR specifies the directory where the source should be"
+  echo "cloned to, and INSTALL_DIR specifies the installation directory."
+  echo "Example:"
+  echo "  build_vim.sh -s ~/devel -t /usr/local"
+  echo "vim will then be cloned to and built in ~/devel/vim, and installed"
+  echo "to /usr/local."
+}
+
+while getopts ":s:t:h" opt; do
+  case ${opt} in
+    s) CLONE_DIR=$OPTARG ;;
+    t) INSTALL_DIR=$OPTARG ;;
+    h) print_help; exit 0 ;;
+    :) echo "Option -$OPTARG requires an argument."; ARGERR=1 ;;
+    \?) echo "Invalid option -$OPTARG"; ARGERR=1 ;;
+  esac
+done
+[[ -z "$CLONE_DIR" ]] && { echo "Missing option -s"; ARGERR=1; }
+[[ -z "$INSTALL_DIR" ]] && { echo "Missing option -t"; ARGERR=1; }
+[[ ! -z "$ARGERR" ]] && { print_help; exit 1; }
+
+set -e
+REPO_DIR=${CLONE_DIR}/vim
+echo "Cloning to ${REPO_DIR}, and installing to ${INSTALL_DIR}..."
+
 # Clone and get to clean slate
 mkdir -p ${CLONE_DIR}
-git -C ${CLONE_DIR} clone https://github.com/vim/vim.git
-REPO_DIR=${CLONE_DIR}/vim
+git -C ${CLONE_DIR} clone https://github.com/vim/vim.git || true
 git -C ${REPO_DIR} clean -fxd
 git -C ${REPO_DIR} pull origin master --rebase
 git -C ${REPO_DIR} checkout master
@@ -47,6 +71,7 @@ cd ${REPO_DIR}
     --enable-gui=gtk2 \
     --enable-cscope \
     --prefix=${INSTALL_DIR}
-make -j${NCPUS} && make install
+make -j${NCPUS}
+make install || { echo "Attempting superuser installation"; sudo make install; }
 
 cd ${CURRENT_DIR}
